requirements.txt
agno
openai
ddgs
pypdf
python-docx
pytesseract
Pillow
python-dotenv
uvicorn
fastapi
python-multipart